# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SubastApp API",
    "description": "API documentation for SubastApp",
    "license": {
      "name": "FabrizioB Website",
      "url": "https://fabriziob.com/"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4002",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/personas/{id}": {
      "get": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "getById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "persona-controller"
        ],
        "operationId": "desactivar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/clientes/{id}/medios-pago/tarjeta": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Agregar tarjeta de crédito",
        "description": "Registra una tarjeta de crédito como medio de pago del cliente",
        "operationId": "addTarjeta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedioPagoTarjetaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tarjeta registrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{id}/medios-pago/cuenta": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Agregar cuenta bancaria",
        "description": "Registra una cuenta bancaria (local o exterior) como medio de pago del cliente",
        "operationId": "addCuenta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedioPagoCuentaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cuenta registrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{id}/medios-pago/cheque": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Agregar cheque",
        "description": "Registra un cheque certificado como medio de pago del cliente",
        "operationId": "addCheque",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedioPagoChequeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cheque registrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
...
    "/productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "summary": "Obtener producto por ID",
        "description": "Devuelve un producto con sus datos completos, extra y lista de IDs de fotos",
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "Productos"
        ],
        "summary": "Actualizar producto",
        "description": "Actualiza parcialmente los campos de un producto. Solo para uso interno de la empresa",
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Campos a actualizar, solo se pisan los que vienen en el body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
...
    "/api/v1/empleados/{id}": {
      "get": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "getById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "desactivar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "empleado-controller"
        ],
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmpleadoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EmpleadoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clientes/{id}/multa": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Asignar multa",
        "description": "Asigna una multa pendiente al cliente. Inhabilita al cliente hasta que sea saldada",
        "operationId": "asignarMulta",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "monto",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Multa asignada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/medios-pago/{mpId}/verificar": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Verificar medio de pago",
        "description": "Marca un medio de pago como verificado por la empresa",
        "operationId": "verificar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "mpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Medio de pago verificado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente o medio de pago no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/medios-pago/{mpId}/desactivar": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Desactivar medio de pago",
        "description": "Desactiva un medio de pago, impidiendo su uso en futuras operaciones",
        "operationId": "desactivar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "mpId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Medio de pago desactivado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MedioPagoResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente o medio de pago no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/inhabilitar": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Inhabilitar cliente",
        "description": "Inhabilita operativamente al cliente para participar en subastas",
        "operationId": "inhabilitar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente inhabilitado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/habilitar": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Habilitar cliente",
        "description": "Habilita operativamente al cliente para participar en subastas",
        "operationId": "habilitar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente habilitado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/categoria": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Actualizar categoría",
        "description": "Cambia la categoría del cliente (comun, especial, plata, oro, platino)",
        "operationId": "actualizarCategoria",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "categoria",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría actualizada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
    "/api/v1/clientes/{id}/admitir": {
      "patch": {
        "tags": [
          "Clientes"
        ],
        "summary": "Admitir cliente",
        "description": "Marca al cliente como admitido para operar en la plataforma",
        "operationId": "admitir",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente admitido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos para acceder a este recurso"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PersonaUpdateRequest": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "foto": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "PersonaResponse": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          }
        }
      },
...
      "MedioPagoTarjetaRequest": {
        "type": "object",
        "properties": {
          "moneda": {
            "$ref": "#/components/schemas/Moneda"
          },
          "titular": {
            "type": "string"
          },
          "ultimos4": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "vencimiento": {
            "type": "string"
          },
          "esInternacional": {
            "type": "boolean"
          }
        }
      },
...
      "MedioPagoCuentaRequest": {
        "type": "object",
        "properties": {
          "moneda": {
            "$ref": "#/components/schemas/Moneda"
          },
          "titular": {
            "type": "string"
          },
          "banco": {
            "type": "string"
          },
          "cbu": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "esExterior": {
            "type": "boolean"
          },
          "iban": {
            "type": "string"
          },
          "pais": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}